requirements.txt
Babel
Flask
Flask-Bootstrap
Flask-Login
Flask-Moment
Flask-PyMongo
Flask-Table
Flask-WTF
pycrypto